vada_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=3.6.2=pyhd8ed1ab_0
aom=3.5.0=h7ea286d_0
appdirs=1.4.4=pyh9f0ad1d_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310h8e9501a_3
asttokens=2.2.1=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
babel=2.11.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.1=pyhd8ed1ab_0
brotli=1.0.9=h1a8c8d9_8
brotli-bin=1.0.9=h1a8c8d9_8
brotlipy=0.7.0=py310h8e9501a_1005
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2023.7.22=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py310h2399d43_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.2=pyhd8ed1ab_0
contourpy=1.0.7=py310h2887b22_0
cryptography=39.0.0=py310hfc83b78_0
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.35=py310h1253130_0
debugpy=1.6.5=py310h0f1eb42_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=py_0
dtw-python=1.3.0=py310h8e9501a_1
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=hb7217d7_0
ffmpeg=5.1.2=gpl_hf318d42_105
filelock=3.10.3=pyhd8ed1ab_0
flit-core=3.8.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.1=h82840c6_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=py310h8e9501a_1
freetype=2.12.1=hd633e50_1
future=0.18.3=pypi_0
gettext=0.21.1=h0186832_0
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.30=pyhd8ed1ab_0
gmp=6.2.1=h9f76cd9_0
gmpy2=2.1.2=py310h2e6cad2_1
gnutls=3.7.8=h9f1a10d_0
h5py=3.7.0=nompi_py310h3419284_102
hdf5=1.12.2=nompi_ha7af310_101
icu=70.1=h6b3803e_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pyha770c72_0
importlib_resources=5.10.2=pyhd8ed1ab_0
ipykernel=6.20.2=pyh736e0ef_0
ipython=8.8.0=pyhd1c38e8_0
ipython_genutils=0.2.0=py_1
jedi=0.18.2=pyhd8ed1ab_0
jellyfish=1.0.1=py310had9acf8_1
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=he4db4b2_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_core=5.1.3=py310hbe9552e_0
jupyter_events=0.6.3=pyhd8ed1ab_0
jupyter_server=2.1.0=pyhd8ed1ab_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyterlab=3.5.2=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.19.0=pyhd8ed1ab_0
kiwisolver=1.4.4=py310h2887b22_1
krb5=1.20.1=h69eda48_0
lame=3.100=h1a8c8d9_1003
lcms2=2.14=h481adae_1
lerc=4.0.0=h9a09cb3_0
libaec=1.0.6=hb7217d7_1
libblas=3.9.0=16_osxarm64_openblas
libbrotlicommon=1.0.9=h1a8c8d9_8
libbrotlidec=1.0.9=h1a8c8d9_8
libbrotlienc=1.0.9=h1a8c8d9_8
libcblas=3.9.0=16_osxarm64_openblas
libcurl=7.87.0=h9049daf_0
libcxx=16.0.4=h4653b0c_0
libdeflate=1.17=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=11_3_0_hd922786_27
libgfortran5=11.3.0=hdaf2cc0_27
libiconv=1.17=he4db4b2_0
libidn2=2.3.4=h1a8c8d9_0
libjpeg-turbo=2.1.4=h1a8c8d9_0
liblapack=3.9.0=16_osxarm64_openblas
libllvm11=11.1.0=hfa12f05_5
libnghttp2=1.51.0=hae82a92_0
libopenblas=0.3.21=openmp_hc731615_3
libpng=1.6.39=h76d750c_0
libprotobuf=3.21.12=hb5ab8b9_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.40.0=h76d750c_0
libssh2=1.10.0=h7a5bd25_3
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.5.0=h5dffbdd_2
libunistring=0.9.10=h3422bc3_0
libvpx=1.11.0=hc470f4d_3
libwebp-base=1.2.4=h57fd34a_0
libxcb=1.13=h9b22ae9_1004
libxml2=2.10.3=h87b0503_0
libzlib=1.2.13=h03a7124_4
llvm-openmp=15.0.7=h7cfbb63_0
llvmlite=0.39.1=py310h1e34944_1
mappy=2.26=pypi_0
markupsafe=2.1.2=py310h8e9501a_0
matplotlib=3.6.3=py310hb6292c7_0
matplotlib-base=3.6.3=py310h78c5c2f_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.4=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.0=he09a6ba_0
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=0.4.8=pyhd8ed1ab_0
nbclient=0.7.2=pyhd8ed1ab_0
nbconvert=7.2.8=pyhd8ed1ab_0
nbconvert-core=7.2.8=pyhd8ed1ab_0
nbconvert-pandoc=7.2.8=pyhd8ed1ab_0
nbformat=5.7.3=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.8.1=h63371fa_1
networkx=3.0=pyhd8ed1ab_0
notebook=6.5.2=pyha770c72_1
notebook-shim=0.2.2=pyhd8ed1ab_0
numba=0.56.4=py310h3124f1e_0
numpy=1.23.5=py310h5d7c261_0
ont-fast5-api=4.1.1=pypi_0
openh264=2.3.1=hb7217d7_1
openjpeg=2.5.0=hbc2ba62_2
openssl=3.1.3=h53f4e23_0
p11-kit=0.24.1=h29577a5_0
packaging=23.0=pyhd8ed1ab_0
pandas=2.0.0=py310h2b830bf_0
pandoc=2.19.2=hce30654_1
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=py_1
patsy=0.5.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310h5f2e7f5_0
pip=23.1.2=pypi_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=2.6.2=pyhd8ed1ab_0
plotly=5.13.1=py_0
pooch=1.6.0=pyhd8ed1ab_0
progressbar33=2.4=py_0
prometheus_client=0.15.0=pyhd8ed1ab_0
prompt-toolkit=3.0.36=pyha770c72_0
protobuf=4.21.12=py310h0f1eb42_0
psutil=5.9.4=py310h8e9501a_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.14.0=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.19.3=py310h8e9501a_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.8=h3ba56d0_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python-json-logger=2.0.4=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytorch=2.0.0=py3.10_0
pytz=2022.7.1=pyhd8ed1ab_0
pyyaml=6.0=py310h8e9501a_5
pyzmq=25.0.0=py310hc407298_0
readline=8.1.2=h46ed386_0
requests=2.28.2=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
scikit-learn=1.2.0=py310ha00a7cd_0
scipy=1.10.0=py310ha0d8a01_0
seaborn=0.12.2=hd8ed1ab_0
seaborn-base=0.12.2=pyhd8ed1ab_0
send2trash=1.8.0=pyhd8ed1ab_0
sentry-sdk=1.15.0=pyhd8ed1ab_0
setproctitle=1.3.2=py310h8e9501a_1
setuptools=67.8.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=3.0.5=pyh44b312d_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=py310ha11ecec_1
svt-av1=1.4.1=h7ea286d_0
sympy=1.11.1=pypyh9d50eac_103
tenacity=8.2.2=pyhd8ed1ab_0
terminado=0.17.1=pyhd1c38e8_0
threadpoolctl=3.1.0=pyh8a188c0_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=2.0.0=py310_cpu
torchinfo=1.7.2=pyhd8ed1ab_0
torchvision=0.15.0=py310_cpu
tornado=6.2=py310h8e9501a_1
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.8.1=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
unicodedata2=15.0.0=py310h8e9501a_0
urllib3=1.26.14=pyhd8ed1ab_0
wandb=0.13.10=pyhd8ed1ab_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=py_1
websocket-client=1.4.2=pyhd8ed1ab_0
wheel=0.38.4=pyhd8ed1ab_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.11.0=pyhd8ed1ab_0
zstd=1.5.2=hf913c23_6
